Requested changes from PR review
Keep a few examples
Correct function signature in gridfunc and pgridfunc
