+++ examples/ex10p.cpp
@@ -180,7 +180,6 @@ int main(int argc, char *argv[])
    double K = 5.0;
    bool visualization = true;
    int vis_steps = 1;
-   bool adios2 = false;
 
    OptionsParser args(argc, argv);
    args.AddOption(&mesh_file, "-m", "--mesh",
@@ -212,10 +211,6 @@ int main(int argc, char *argv[])
                   "Enable or disable GLVis visualization.");
    args.AddOption(&vis_steps, "-vs", "--visualization-steps",
                   "Visualize every n-th timestep.");
-   args.AddOption(&adios2, "-adios2", "--adios2-streams", "-no-adios2",
-                  "--no-adios2-streams",
-                  "Save data adios2 streams, files can use ParaView (paraview.org) VTX visualization.");
-
    args.Parse();
    if (!args.Good())
    {
@@ -356,25 +351,6 @@ int main(int argc, char *argv[])
       }
    }
 
-   // 10. Optionally output a BP (binary pack file) ADIOS2DataCollection
-   //     ADIOS2: https://adios2.readthedocs.io
-#ifdef MFEM_USE_ADIOS2
-   ADIOS2DataCollection* adios2_dc = NULL;
-
-   if (adios2)
-   {
-      std::string postfix(mesh_file);
-      postfix.erase(0, std::string("../data/").size() );
-      postfix += "_solver" + std::to_string(ode_solver_type);
-      const std::string collection_name = "ex10-p-" + postfix + ".bp";
-
-      adios2_dc = new ADIOS2DataCollection(MPI_COMM_WORLD, collection_name, pmesh);
-      adios2_dc->SetParameter("SubStreams", std::to_string(num_procs/2) );
-      adios2_dc->RegisterField("velocity", &v_gf);
-      adios2_dc->RegisterField("elastic_energy", &w_gf);
-   }
-#endif
-
    double ee0 = oper.ElasticEnergy(x_gf);
    double ke0 = oper.KineticEnergy(v_gf);
    if (myid == 0)
@@ -388,7 +364,7 @@ int main(int argc, char *argv[])
    oper.SetTime(t);
    ode_solver->Init(oper);
 
-   // 11. Perform time-integration
+   // 10. Perform time-integration
    //     (looping over the time iterations, ti, with a time-step dt).
    bool last_step = false;
    for (int ti = 1; !last_step; ti++)
@@ -421,26 +397,10 @@ int main(int argc, char *argv[])
                visualize(vis_w, pmesh, &x_gf, &w_gf);
             }
          }
-
-#ifdef MFEM_USE_ADIOS2
-         if (adios2)
-         {
-            adios2_dc->SetTime(t);
-            adios2_dc->SetCycle(ti);
-            adios2_dc->Save();
-         }
-#endif
       }
    }
 
-#ifdef MFEM_USE_ADIOS2
-   if (adios2)
-   {
-      delete adios2_dc;
-   }
-#endif
-
-   // 12. Save the displaced mesh, the velocity and elastic energy.
+   // 11. Save the displaced mesh, the velocity and elastic energy.
    {
       v_gf.SetFromTrueVector(); x_gf.SetFromTrueVector();
       GridFunction *nodes = &x_gf;
@@ -465,7 +425,7 @@ int main(int argc, char *argv[])
       w_gf.Save(ee_ofs);
    }
 
-   // 13. Free the used memory.
+   // 12. Free the used memory.
    delete ode_solver;
    delete pmesh;
