+++ examples/ex4p.cpp
@@ -61,7 +61,6 @@ int main(int argc, char *argv[])
    bool static_cond = false;
    bool hybridization = false;
    bool visualization = 1;
-   bool adios2 = false;
 
    OptionsParser args(argc, argv);
    args.AddOption(&mesh_file, "-m", "--mesh",
@@ -79,10 +78,6 @@ int main(int argc, char *argv[])
    args.AddOption(&visualization, "-vis", "--visualization", "-no-vis",
                   "--no-visualization",
                   "Enable or disable GLVis visualization.");
-   args.AddOption(&adios2, "-adios2", "--adios2-streams", "-no-adios2",
-                  "--no-adios2-streams",
-                  "Save data adios2 streams, files can use ParaView (paraview.org) VTX reader visualization.");
-
    args.Parse();
    if (!args.Good())
    {
@@ -274,32 +269,7 @@ int main(int argc, char *argv[])
       sol_sock << "solution\n" << *pmesh << x << flush;
    }
 
-   // 17. Optionally output a BP (binary pack file) ADIOS2DataCollection
-   //     ADIOS2: https://adios2.readthedocs.io
-#ifdef MFEM_USE_ADIOS2
-   if (adios2)
-   {
-      std::string postfix(mesh_file);
-      postfix.erase(0, std::string("../data/").size() );
-      postfix += "_o" + std::to_string(order);
-      // use extension .bp as it's recognized by ParaView
-      // output is a bp directory by default
-      const std::string collection_name = "ex4-p_" + postfix + ".bp";
-
-      // Create adios2 data collection, requires the communicator
-      ADIOS2DataCollection adios2_dc(MPI_COMM_WORLD, collection_name, pmesh);
-      // for other parameters of interest:
-      // https://adios2.readthedocs.io/en/latest/engines/engines.html#bp4
-      // For example:
-      // sets number of data streams inside BP directory, appropriate for running at scale
-      adios2_dc.SetParameter("SubStreams", std::to_string(num_procs/2) );
-      //adios2_dc.SetLevelsOfDetail(2);
-      adios2_dc.RegisterField("sol", &x);
-      adios2_dc.Save();
-   }
-#endif
-
-   // 18. Free the used memory.
+   // 17. Free the used memory.
    delete pcg;
    delete prec;
    delete hfes;
